+++ prettier.config.cjs
@@ -1,4 +1,4 @@
-/** @type {import("prettier").Config} */
-module.exports = {
-  plugins: [require.resolve("prettier-plugin-tailwindcss")],
-};
+// /** @type {import("prettier").Config} */
+// module.exports = {
+//   plugins: [require.resolve("prettier-plugin-tailwindcss")],
+// };
